DONE: Segregate user routes
