+++ server/views/login.handlebars
@@ -3,7 +3,7 @@
 </header>
 
 <main>
-    <form action="/login" method="POST">
+    <form action="/users/login" method="POST">
     <div class="form-group mb-3">
         <label>Username</label>
         <input type="text" class="form-control" placeholder="Username" name="username">
