+++ server/views/profile.handlebars
@@ -1,1 +1,6 @@
 <h1>Hi {{user}}</h1>
+<form action="/users/logout" method="POST">
+    <div>
+    <button type="submit" class="btn btn-primary form-control">Log Out</button>
+    </div>
+</form>
