+++ server/views/register.handlebars
@@ -3,14 +3,10 @@
 </header>
 
 <main>
-    <form action="/register" method="POST">
+    <form action="/users/register" method="POST">
     <div class="form-group mb-3">
         <label>Username</label>
         <input type="text" class="form-control" placeholder="Username" name="username">
-    </div>
-    <div class="form-group mb-3">
-        <label>Email</label>
-        <input type="text" class="form-control" placeholder="Email" name="email">
     </div>
     <div class="form-group mb-3">
         <label>Password</label>
